--- basic/blockchain.js
@@ -1,18 +1,0 @@
-// 블록의 생성, 검증, 합의 알고리즘
-// 프로토콜 변경 시 수정
-
-class Block {
-  constructor(header, body) {
-    this.header = header
-    this.body = body
-  }
-}
-
-class BlockHeader{
-  constructor() {
-    
-  }
-}
-
-
-const blockchain = []
