--- basic/main.js
@@ -1,10 +1,0 @@
-// HTTP Server / P2P Server / 지갑 초기화
-// (사용자와 노드 간의 통신)
-
-function initHttpServer(){
-
-}
-
-initHttpServer();
-initP2PServer();
-initWallet();
