--- basic/utils.js
@@ -1,13 +1,0 @@
-// 도구적인 기능을 하는 함수들 포함
-
-function getCurrentTimestamp() {
-
-}
-
-function getCurrentVersion() {
-
-}
-
-function hexToBinary(s) {
-  
-}
