--- basic/wallet.js
@@ -1,13 +1,0 @@
-// 비밀키와 공개키를 관리
-/* 
-  지갑에서부터 비밀키를 가져오거나 공개키를 생성
-  새로운 비밀키를 생서하거나 있으면 기존 파일에서 읽어오는 기능
-*/
-
-function generatePrivateKey() {
-
-}
-
-function initWallet() {
-  
-}
