+++ networks.js
@@ -1,12 +1,8 @@
 // P2P서버 관련 내용 - 메세지 핸들러, 에러 핸들러, 브로드캐스트 등
 // (노드와 노드 간의 통신)
 
-const MessageType = {
-
-}
+const MessageType = {};
 
 const sockets = [];
 
-function initP2PServer() {
-  
-}
+function initP2PServer() {}
