+++ package.json
@@ -1,23 +1,16 @@
 {
-  "dependencies": {
-    "body-parser": "^1.19.1",
-    "crypto": "^1.0.1",
-    "crypto-js": "^4.1.1",
-    "elliptic": "^6.5.4",
-    "express": "^4.17.2",
-    "fs": "^0.0.1-security",
-    "merkle": "^0.6.0",
-    "nodejs-websocket": "^1.7.2",
-    "random": "^3.0.6",
-    "ws": "^8.4.0"
-  },
-  "name": "user4",
+  "name": "basic",
   "version": "1.0.0",
-  "main": "block.js",
+  "description": "",
+  "main": "main.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
-  "author": "jeonjinyoung",
+  "author": "",
   "license": "ISC",
-  "description": ""
+  "dependencies": {
+    "crypto-js": "^4.1.1",
+    "cryptojs": "^2.5.3",
+    "merkle": "^0.6.0"
+  }
 }
