+++ resson/restartServer.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+kill -9 `ps -ef | grep httpServer.js | grep node | awk '{print $2}'`
+node main.js &
