+++ utils/u_hash.js
@@ -1,0 +1,25 @@
+const cryptojs = require("crypto-js");
+
+function createHash(data) {
+  const {
+    version,
+    index,
+    previousHash,
+    timestamp,
+    merkleRoot,
+    difficulty,
+    nonce,
+  } = data.header;
+  const blockString =
+    version +
+    index +
+    previousHash +
+    timestamp +
+    merkleRoot +
+    difficulty +
+    nonce;
+  const hash = cryptojs.SHA256(blockString).toString();
+  return hash;
+}
+
+module.exports = { createHash };
